+++ docker/package/tezos_setup_wizard.py
@@ -9,13 +9,14 @@ A wizard utility to help set up tezos-baker.
 Asks questions, validates answers, and executes the appropriate steps using the final configuration.
 """
 
-import os, sys, subprocess, shlex, shutil
+import os, sys, shutil
 import readline
-import re, textwrap
+import re
 import urllib.request
 import json
 from typing import List
 
+from wizard_structure import *
 
 # Global options
 
@@ -59,6 +60,7 @@ history_modes = {
 
 TMP_SNAPSHOT_LOCATION = "/tmp/tezos_node.snapshot"
 
+
 # Regexes
 
 ledger_regex = b"ledger:\/\/[\w\-]+\/[\w\-]+\/[\w']+\/[\w']+"
@@ -66,122 +68,24 @@ secret_key_regex = b"(encrypted|unencrypted):(?:\w{54}|\w{88})"
 address_regex = b"tz[123]\w{33}"
 derivation_path_regex = b"(?:bip25519|ed25519|secp256k1|P-256)\/[0-9]+h\/[0-9]+h"
 
-# Input validators
-
-
-def enum_range_validator(options):
-    def _validator(input):
-        intrange = list(map(str, range(1, len(options) + 1)))
-        if input not in intrange and input not in options:
-            raise ValueError(
-                "Please choose one of the provided values or use their respective numbers: "
-                + ", ".join(options)
-            )
-        try:
-            opt = int(input) - 1
-        except:
-            return input
-        else:
-            opts = options
-            if isinstance(options, dict):
-                opts = list(options.keys())
-            return opts[opt]
-
-    return _validator
-
-
-def filepath_validator(input):
-    if input and not os.path.isfile(input):
-        raise ValueError("Please input a valid file path.")
-    return input
-
-
-def required_field_validator(input):
-    if not input.strip():
-        raise ValueError("Please provide this required option.")
-    return input
-
-
-# The input has to be valid to at least one of the two passed validators.
-def or_validator(validator1, validator2):
-    def _validator(input):
-        try:
-            return validator1(input)
-        except:
-            return validator2(input)
-
-    return _validator
-
-
-# Runs the input through the passed validator, allowing for possible alteration,
-# but doesn't raise an exception if it doesn't validate to allow for custom options, too.
-def or_custom_validator(validator):
-    def _validator(input):
-        result = input
-        try:
-            result = validator(input)
-        except:
-            pass
-        return result
-
-    return _validator
-
-
-# To be validated, the input should adhere to the Tezos secret key format:
-# {encrypted, unencrypted}:<base58 encoded string with length 54 or 88>
-def secret_key_validator(input):
-    match = re.match(secret_key_regex.decode("utf-8"), input.strip())
-    if not bool(match):
-        raise ValueError(
-            "The input doesn't match the format for the Tezos secret key: "
-            "{{encrypted, unencrypted}:<base58 encoded string with length 54 or 88>}"
-            "\nPlease check the input and try again."
-        )
-    return input
-
-
-# To be validated, the input should adhere to the derivation path format:
-# [0-9]+h/[0-9]+h
-def derivation_path_validator(input):
-    match = re.match(b"[0-9]+h\/[0-9]+h".decode("utf-8"), input.strip())
-    if not bool(match):
-        raise ValueError(
-            "The input doesn't match the format for the derivation path: "
-            "'[0-9]+h/[0-9]+h'"
-            "\nPlease check the input and try again."
-        )
-    return input
-
-
-class Validator:
-    def __init__(self, validator):
-        self.validator = validator
-
-    def validate(self, input):
-        if self.validator is not None:
-            if isinstance(self.validator, list):
-                for v in self.validator:
-                    input = v(input)
-                return input
-            else:
-                return self.validator(input)
-
 
 # Wizard CLI utility
 
 
-suppress_warning_text = "TEZOS_CLIENT_UNSAFE_DISABLE_DISCLAIMER=YES"
+welcome_text = """Tezos Setup Wizard
 
+Welcome, this wizard will help you to set up the infrastructure to interact with the
+Tezos blockchain.
 
-def proc_call(cmd):
-    return subprocess.check_call(shlex.split(cmd))
+In order to run a baking instance, you'll need the following Tezos binaries:
+ tezos-client, tezos-node, tezos-baker-<proto>, tezos-endorser-<proto>.
+If you have installed tezos-baking, these binaries are already installed.
 
+All commands within the service are run under the 'tezos' user.
 
-def get_proc_output(cmd):
-    if sys.version_info.major == 3 and sys.version_info.minor < 7:
-        return subprocess.run(shlex.split(cmd), stdout=subprocess.PIPE)
-    else:
-        return subprocess.run(shlex.split(cmd), capture_output=True)
+To access help and possible options for each question, type in 'help' or '?'.
+Type in 'exit' to quit.
+"""
 
 
 def fetch_snapshot(url):
@@ -190,32 +94,6 @@ def fetch_snapshot(url):
     urllib.request.urlretrieve(url, filename, progressbar_hook)
     print()
     return filename
-
-
-def progressbar_hook(chunk_number, chunk_size, total_size):
-    done = chunk_number * chunk_size
-    percent = min(int(done * 100 / total_size), 100)
-    print("Progress:", percent, "%,", int(done / (1024 * 1024)), "MB", end="\r")
-
-
-def color(input, colorcode):
-    return colorcode + input + "\x1b[0m"
-
-
-def yes_or_no(prompt, default=None):
-    valid = False
-    while not valid:
-        answer = input(prompt).strip().lower()
-        if not answer and default is not None:
-            answer = default
-        if answer in ["y", "yes"]:
-            print()
-            return True
-        elif answer in ["n", "no"]:
-            print()
-            return False
-        else:
-            print(color("Please provide a 'yes' or 'no' answer.", "\x1b[1;31m"))
 
 
 def wait_for_ledger_baking_app():
@@ -235,24 +113,6 @@ def wait_for_ledger_baking_app():
         )
         ledgers_derivations.setdefault(ledger_url, []).append(derivation_path)
     return ledgers_derivations
-
-
-def get_data_dir(network):
-    output = get_proc_output("systemctl show tezos-node-" + network + ".service").stdout
-    config = re.search(b"Environment=(.*)(?:$|\n)", output)
-    if config is None:
-        print(
-            "tezos-node-" + network + ".service configuration not found, "
-            "defaulting to /var/lib/tezos/node-" + network
-        )
-        return "/var/lib/tezos/node-" + network
-    config = config.group(1)
-    data_dir = re.search(b"DATA_DIR=(.*?)(?: |$|\n)", config)
-    if data_dir is not None:
-        return data_dir.group(1).decode("utf-8")
-    else:
-        print("DATA_DIR is undefined, defaulting to /var/lib/tezos/node-" + network)
-        return "/var/lib/tezos/node-" + network
 
 
 def ledger_urls_info(ledgers_derivations, node_endpoint):
@@ -283,14 +143,6 @@ def ledger_urls_info(ledgers_derivations, node_endpoint):
     return ledgers_info
 
 
-def search_baking_service_config(config_contents, regex, default):
-    res = re.search(regex, config_contents)
-    if res is None:
-        return default
-    else:
-        return res.group(1)
-
-
 def is_full_snapshot(import_mode):
     if import_mode == "download full":
         return True
@@ -300,79 +152,6 @@ def is_full_snapshot(import_mode):
         ).stdout
         return re.search(b"at level [0-9]+ in full", output) is not None
     return False
-
-
-class Step:
-    def __init__(
-        self,
-        id: str,
-        prompt: str,
-        help: str,
-        default: str = "1",
-        options=None,
-        validator=None,
-    ):
-        self.id = id
-        self.prompt = prompt
-        self.help = help
-        self.default = default
-        self.options = options
-        self.validator = validator
-
-    def pprint_options(self):
-        i = 1
-        def_i = None
-        try:
-            def_i = int(self.default)
-        except:
-            pass
-
-        if self.options and isinstance(self.options, list):
-            options_count = 0
-            for o in self.options:
-                if isinstance(o, dict):
-                    for values in o.values():
-                        if not isinstance(values, list):
-                            options_count += 1
-                        else:
-                            options_count += len(values)
-                else:
-                    options_count += 1
-            index_len = len(str(options_count))
-            str_format = f"{{:{index_len}}}. {{}}"
-            for o in self.options:
-                if isinstance(o, dict):
-                    for k, values in o.items():
-                        print()
-                        print(f"'{k}':")
-                        print()
-                        if not isinstance(values, list):
-                            values = [values]
-                        for v in values:
-                            if def_i is not None and i == def_i:
-                                print(str_format.format(i, "(default) " + v))
-                            else:
-                                print(str_format.format(i, v))
-                            i += 1
-                    print()
-                else:
-                    if def_i is not None and i == def_i:
-                        print(str_format.format(i, "(default) " + o))
-                    else:
-                        print(str_format.format(i, o))
-                    i += 1
-        elif self.options and isinstance(self.options, dict):
-            index_len = len(str(len(self.options)))
-            str_format = f"{{:{index_len}}}. {{:<26}}  {{}}"
-            for o in self.options:
-                description = textwrap.indent(
-                    textwrap.fill(self.options[o], 60), " " * 31
-                ).lstrip()
-                if def_i is not None and i == def_i:
-                    print(str_format.format(i, o + " (default)", description))
-                else:
-                    print(str_format.format(i, o, description))
-                i += 1
 
 
 # Steps
@@ -533,73 +312,7 @@ def get_ledger_derivation(ledgers_derivations, node_endpoint):
     )
 
 
-class Setup:
-    def __init__(self, config={}):
-        self.config = config
-
-    def query_step(self, step: Step):
-
-        validated = False
-        while not validated:
-            print(step.prompt)
-            step.pprint_options()
-            answer = input("> ").strip()
-
-            if answer.lower() in ["quit", "exit"]:
-                raise KeyboardInterrupt
-            elif answer.lower() in ["help", "?"]:
-                print(step.help)
-                print()
-            else:
-                if not answer and step.default is not None:
-                    answer = step.default
-
-                try:
-                    if step.validator is not None:
-                        answer = step.validator.validate(answer)
-                except ValueError as e:
-                    print(color("Validation error: " + str(e), "\x1b[1;31m"))
-                else:
-                    validated = True
-                    self.config[step.id] = answer
-
-    def fill_baking_config(self):
-        network = self.config["network"]
-        output = get_proc_output(
-            "systemctl show tezos-baking-" + network + ".service"
-        ).stdout
-        config_filepath = re.search(b"EnvironmentFiles=(.*) ", output)
-        if config_filepath is None:
-            print(
-                "EnvironmentFiles not found in tezos-baking-"
-                + network
-                + ".service configuration,",
-                "defaulting to /etc/default/tezos-baking-" + network,
-            )
-            config_filepath = "/etc/default/tezos-baking-" + network
-        else:
-            config_filepath = config_filepath.group(1).decode().strip()
-
-        with open(config_filepath, "r") as f:
-            config_contents = f.read()
-            self.config["client_data_dir"] = search_baking_service_config(
-                config_contents, 'DATA_DIR="(.*)"', "/var/lib/.tezos-client"
-            )
-            self.config["node_rpc_addr"] = search_baking_service_config(
-                config_contents, 'NODE_RPC_ENDPOINT="(.*)"', "http://localhost:8732"
-            )
-            self.config["baker_alias"] = search_baking_service_config(
-                config_contents, 'BAKER_ADDRESS_ALIAS="(.*)"', "baker"
-            )
-
-    def get_tezos_client_options(self):
-        return (
-            "--base-dir "
-            + self.config["client_data_dir"]
-            + " --endpoint "
-            + self.config["node_rpc_addr"]
-        )
-
+class Setup(Setup):
     def get_current_head_level(self):
         response = urllib.request.urlopen(
             self.config["node_rpc_addr"] + "/chains/main/blocks/head/header"
@@ -619,7 +332,9 @@ class Setup:
             print()
             if yes_or_no("Delete this data and bootstrap the node again? <y/N> ", "no"):
                 for path in diff:
-                    if proc_call("sudo rm -r " + os.path.join(node_dir, path)):
+                    try:
+                        proc_call("sudo rm -r " + os.path.join(node_dir, path))
+                    except:
                         print(
                             "Could not clean the Tezos node data directory. "
                             "Please do so manually."
@@ -930,52 +645,9 @@ class Setup:
     def start_baking(self):
         self.systemctl_start_action("baking")
 
-    def systemctl_start_action(self, service):
-        proc_call(
-            "sudo systemctl start tezos-"
-            + service
-            + "-"
-            + self.config["network"]
-            + ".service"
-        )
-
-    def systemctl_enable(self):
-        if self.config["systemd_mode"] == "yes":
-            print(
-                "Enabling the tezos-{}-{}.service".format(
-                    self.config["mode"], self.config["network"]
-                )
-            )
-            proc_call(
-                "sudo systemctl enable tezos-"
-                + self.config["mode"]
-                + "-"
-                + self.config["network"]
-                + ".service"
-            )
-        else:
-            print("The services won't restart on boot.")
-
     def run_setup(self):
 
-        print("Tezos Setup Wizard")
-        print()
-        print(
-            "Welcome, this wizard will help you to set up the infrastructure",
-            "to interact with the Tezos blockchain.",
-        )
-        print(
-            "In order to run a baking instance, you'll need the following Tezos binaries:\n",
-            "tezos-client, tezos-node, tezos-baker-<proto>, tezos-endorser-<proto>.\n",
-            "If you have installed tezos-baking, these binaries are already installed.",
-        )
-        print("All commands within the service are run under the 'tezos' user.")
-        print()
-        print(
-            "To access help and possible options for each question, type in 'help' or '?'."
-        )
-        print("Type in 'exit' to quit.")
-        print()
+        print(welcome_text)
 
         self.query_step(network_query)
         self.fill_baking_config()
